Заменено 9 задание
