--- hw09_generator_of_validators/go-validate/main.go
@@ -1,5 +1,0 @@
-package main
-
-func main() {
-	// Place your code here
-}
