--- hw09_generator_of_validators/go.mod
@@ -1,5 +1,0 @@
-module github.com/fixme_my_friend/hw09_generator_of_validators
-
-go 1.14
-
-require github.com/stretchr/testify v1.5.1
