+++ hw09_struct_validator/validator.go
@@ -1,0 +1,17 @@
+package hw09_struct_validator //nolint:golint,stylecheck
+
+type ValidationError struct {
+	Field string
+	Err   error
+}
+
+type ValidationErrors []ValidationError
+
+func (v ValidationErrors) Error() string {
+	panic("implement me")
+}
+
+func Validate(v interface{}) error {
+	// Place your code here
+	return nil
+}
